pyproject.toml
[project]
name = "taiwan-holidays"
version = "0.2026.0"
description = "A Python package for determining whether a given date is a workday or a holiday in Taiwan, based on the official administrative calendar."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "chesham", email = "c.moweb@gmail.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "pandas==2.2.3",
    "python-dateutil==2.9.0.post0",
]

[project.urls]
homepage = "https://github.com/Chesham/taiwan-holidays"
repository = "https://github.com/Chesham/taiwan-holidays"
documentation = "https://github.com/Chesham/taiwan-holidays"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "behave==1.2.6",
    "black==24.10.0",
    "coverage==7.6.9",
    "grappa==1.0.1",
    "tox>=4.30.2",
]
